requirements.txt
streamlit==1.32.0
google-generativeai==0.3.1
python-dotenv==1.0.0
fpdf2==2.7.8
pillow==9.0.1
pandas==1.5.3
matplotlib==3.7.2